src/ingestion-dispatcher/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-pubsub
pypdf
cloudevents